requirements.txt
asgiref==3.8.1
Django==4.2
django-admin-interface==0.29.2
django-colorfield==0.11.0
django-tinymce==3.6.1
django-versatileimagefield==3.0
pillow==11.1.0
python-decouple==3.8
python-magic==0.4.27
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
